Renamed BeiMock pipeline to mock.illumina
