+++ src/test/scala/mg7/mock/illumina.scala
@@ -1,8 +1,9 @@
-package ohnosequences.test.mg7
+package ohnosequences.test.mg7.mock
 
-import ohnosequences.test.mg7.testDefaults._
+import ohnosequences.test.mg7._, testDefaults._
 import ohnosequences.mg7._, loquats._
-import ohnosequences.datasets._, illumina._
+import ohnosequences.datasets._
+import ohnosequences.datasets.illumina._
 import ohnosequences.cosas._, types._, klists._
 import ohnosequences.loquat._
 import ohnosequences.statika._, aws._
@@ -13,9 +14,9 @@ import ohnosequences.awstools.autoscaling._
 import ohnosequences.awstools.regions.Region._
 import com.amazonaws.auth._, profile._
 
-case object BeiMock {
+case object illumina {
 
-  case object pipeline extends FlashMG7Pipeline(testDefaults.IlluminaParameters) with MG7PipelineDefaults {
+  case object pipeline extends FlashMG7Pipeline(IlluminaParameters) with MG7PipelineDefaults {
 
     // TODO move all this to the testData object
     /* For now we are only testing one sample */
@@ -37,7 +38,7 @@ case object BeiMock {
 
     val outputS3Folder = testDefaults.outputS3FolderFor("illumina")
 
-    val flashParameters = FlashParameters(illumina.bp250)
+    val flashParameters = FlashParameters(bp250)
 
     val flashConfig: AnyFlashConfig = FlashConfig(1)
   }
